Mic Selection Improvements
Mic selection is now done through the icon. All mics are listed even if their internal system label doesn't start with M. Different mics can be select. 

Still there is a bug that needs resolved. The selection of a different mic only works if you go and select that microphone as the system default in control panel